schedule.rsの内容をlib.rsへ移した
